+++ adbnik/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "1.4.2"
+__version__ = "2.0.0"
 APP_TITLE = "Adbnik"